breaking videoEmbedCleaning into its own files
